Correct unnecessary boxing of enums
The code generator attempts to find cycles in the protobuf specification
by recursively marking nodes that are not part of a cycle as leaf nodes,
and then inserting Box<> here-and-there until no more cycles exists. But
the function forgets to consider enums as leaf nodes, causing
unnecessary boxing and in some cases broken code.