Replace from_reader convention for message reading with a MessageRead trait.
This has the advantage of not needing to pass the messages's TypeName:from_reader function,
but the disadvantage of needing to specify reader.read_message::<TypeName>(...) using a turbofish
in all cases where the type can't be inferred.

This also may have the disadvantage of removing the ability to specify a custom read function for
a message type, but I'm unsure how common of a practice that is.