Add possibility to read message with speciefied length (#109)
* Add possibility to read message with speciefied length

As you can read in https://developers.google.com/protocol-buffers/docs/techniques
the protobuf protocol does not specify the format how the length of a message
is stated. If you communicate with an entity which does not encode it's message
length in varints you are not able to read them. Therfore this fix should enable
you to read your own format of message size and read the message of specified
length.

* Remove println statement

Was used during debuging and I forgot to delete it earlier.

* Rename read_len(_size) and read_message(_size)

Add documentation.